tests/wpt/web-platform-tests/payment-request/payment-request-update-event-constructor.http.html

<!DOCTYPE html>
<!-- Copyright © 2017 Chromium authors and World Wide Web Consortium, (Massachusetts Institute of Technology, ERCIM, Keio University, Beihang). -->
<meta charset="utf-8">
<title>Test for PaymentRequestUpdateEvent Constructor</title>
<link rel="help" href="https://w3c.github.io/browser-payment-api/#constructor">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
  assert_throws("SecurityError", () => {
    new PaymentRequestUpdateEvent("test");
  });
}, "PaymentRequestUpdateEvent constructor throws in a non-secure context");
</script>
